src/cement_operations_optimization/data_generator/requirements.txt
flask-socketio>=5.5.1
flask>=3.1.2
psycopg2-binary>=2.9.10
google-genai>=1.32.0
google-cloud-bigquery>=3.36.0
firebase-admin>=7.1.0
google-cloud-storage>=3.3.1
google-cloud-vision>=3.10.2
fastapi>=0.110.0
uvicorn>=0.29.0
pandas>=2.3.2
python-dotenv>=1.0.0
dotenv>=0.9.9
google-cloud-pubsub>=2.31.1